code using ZIO-RC17

+++ src/main/scala/com/zio/examples/http4s_doobie/Data.scala
@@ -1,5 +1,5 @@
 package com.zio.examples.http4s_doobie
 
-final case class User(id: Int, name: String)
+final case class User(id: Long, name: String)
 
 final case class UserNotFound(id: Int) extends Exception
